move comparison::Metric to database crate
