+++ site/src/request_handlers/self_profile.rs
@@ -4,13 +4,12 @@ use std::sync::Arc;
 use std::time::Instant;
 
 use bytes::Buf;
-use database::CommitType;
+use database::{comparison::Metric, CommitType};
 use headers::{ContentType, Header};
 use hyper::StatusCode;
 
 use crate::api::self_profile::ArtifactSizeDelta;
 use crate::api::{self_profile, self_profile_processed, self_profile_raw, ServerResult};
-use crate::comparison::Metric;
 use crate::db::ArtifactId;
 use crate::load::SiteCtxt;
 use crate::selector::{self};
